H3: transferencia

+++ src/main/java/com/keepu/webAPI/dto/response/TransactionResponse.java
@@ -3,10 +3,11 @@ package com.keepu.webAPI.dto.response;
 import com.keepu.webAPI.model.enums.StoreType;
 import com.keepu.webAPI.model.enums.TransactionType;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public record TransactionResponse(
-        Double amount,
+        BigDecimal amount,
         String description,
         TransactionType transactionType,
         LocalDateTime transactionDate,
